use SLURM_JOB_ID; echo job description before building; allow re-using sandbox

+++ job.sh
@@ -13,17 +13,23 @@ set -euo pipefail
 mkdir -p /tmp/nyx/{run,sandbox,job}
 export XDG_RUNTIME_DIR=/tmp/nyx/run
 
-[[ -e /tmp/nyx/sandbox ]] && podman unshare rm -rf /tmp/nyx/sandbox
-singularity build --sandbox /tmp/nyx/sandbox $HOME/nyx/guest/latest
+if [[ "${2:-}" != "-e" ]]; then
+  [[ -e /tmp/nyx/sandbox ]] && podman unshare rm -rf /tmp/nyx/sandbox
+  singularity build --sandbox /tmp/nyx/sandbox $HOME/nyx/guest/latest
+fi
 
-_NYX_CURRENT="/tmp/nyx/job/$(date +%Y%m%d%H%S)"
+_NYX_CURRENT="/tmp/nyx/job/$SLURM_JOB_ID"
 mkdir "$_NYX_CURRENT"
+
+_NYX_TARGET_FLAKE="github:chaotic-cx/nyx${1:-/bump/$(date +%Y%m%d)-1}"
+
+echo "Building '$_NYX_TARGET_FLAKE' job $SLURM_JOB_ID at $(hostname)"
 
 exec singularity exec --writable --fakeroot --no-home --containall \
   -B '/dev/full:/dev/full' \
   -B "$_NYX_CURRENT:/tmp/nyx-wd" \
   --workdir /tmp /tmp/nyx/sandbox \
-  nix develop "github:chaotic-cx/nyx${1:-/bump/$(date +%Y%m%d)-1}" -c env \
+  nix develop "$_NYX_TARGET_FLAKE" -c env \
   NYX_WD="/tmp/nyx-wd" \
   CACHIX_AUTH_TOKEN="$(cat $HOME/nyx/cachix.secret)" \
   chaotic-nyx-build
